Fix drag-n-drop for local files
Looking at the comments on DroppedFile [1] and the egui source [2]
it looks like getting a path vs. bytes is mutually exclusive, depending
on platform.
So use either which one is available, preferring the bytes that would
already be loaded into memory.

The "drawback" of this implementation is that we'd "forget" that
a local file was originally dragged, showing a different text in the
info bar.

[1] https://github.com/emilk/egui/blob/68b3ef7f6badfe893d3bbb1f791b481069d807d9/crates/egui/src/data/input.rs#L325
[2] https://github.com/emilk/egui/blob/68b3ef7f6badfe893d3bbb1f791b481069d807d9/crates/egui-winit/src/lib.rs#L407